+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Non-function properties will now no longer be shown by default when autocompleting a method call (e.g., `foo:bar`).
   To revert back to the original behaviour, enable `luau-lsp.completion.showPropertiesOnMethodCall`
 
+### Fixed
+
+- Autocompletion of variables that hold a class type will now correctly have a kind of "variable" rather than "class"
+
 ## [1.27.1] - 2024-01-20
 
 ### Changed